+++ Laptop/Models/Order.cs
@@ -18,7 +18,7 @@ namespace Laptop.Models
         public bool Status { get; set; }
         public string? Voucher { get; set; }
 
-        public int ? Total { get; set; }
+        public long ? Total { get; set; }
 
         public virtual AppUserModel Customer { get; set; } = null!;
         public virtual Voucher? VoucherNavigation { get; set; }
