+++ Laptop/Service/IMomoService.cs
@@ -5,7 +5,7 @@ namespace Laptop.Service
 {
     public interface IMomoService
     {
-        Task<MomoCreatePaymentResponseModel> CreatePaymentAsync(int total);
+        Task<MomoCreatePaymentResponseModel> CreatePaymentAsync(long total);
         MomoExecuteResponseModel PaymentExecuteAsync(IQueryCollection collection);
     }
 }
